backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
anthropic>=0.86.0
python-dotenv==1.0.1
pydantic==2.7.3
supabase>=2.0.0
tavily-python>=0.7.0
httpx>=0.27.0
openpyxl==3.1.4
